Handel Appointly v1.0

+++ src/main/java/com/handel/HandelAppointly/repositorios/MedicinaRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MedicinaRepositorio extends JpaRepository<Medicina, Integer> {
+public interface MedicinaRepositorio extends JpaRepository<Medicina, Long> {
 }
